requirements.txt
appdirs==1.4.4
asgiref==3.3.4
colorama==0.4.4
distlib==0.3.1
Django==3.2.3
django-phonenumber-field==5.1.0
djangorestframework==3.12.4
filelock==3.0.12
importlib-metadata==4.0.1
joblib==1.0.1
numpy==1.20.3
packaging==20.9
phonenumberslite==8.12.23
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytz==2021.1
scikit-learn==0.24.2
scipy==1.6.3
six==1.16.0
sklearn==0.0
sqlparse==0.4.1
threadpoolctl==2.1.0
toml==0.10.2
tox==3.23.1
typing-extensions==3.10.0.0
virtualenv==20.4.6
zipp==3.4.1
